Refactor: Added screens instead of components
