+++ src/pages/Archives.jsx
@@ -3,7 +3,8 @@ import { useAuth } from "../contexts/providers/AuthProvider";
 import { useUserData } from "../contexts/providers/userDataProvider";
 import { fetchArchives } from "../services";
 import { MdArchive } from "../utils/icons";
-import NoteCard from "./NoteCard";
+import { NoteCard } from "../components";
+import { useDocumentTitle } from "../Hooks";
 
 function Archives() {
   const {
@@ -14,6 +15,8 @@ function Archives() {
   const {
     authState: { token },
   } = useAuth();
+
+  useDocumentTitle("Archives");
 
   useEffect(() => {
     fetchArchives(token, userDataDispatch);
